add(sequence.c) + fix(Makefile): new file + compilation error fix
The compilation was crashing without -fsanitize because the math dynamic library was missing: -lm
