+++ Makefile
@@ -9,9 +9,9 @@ OBJ_DIR := objs
 SRC_DIR := srcs
 INC_DIR := inc
 
-CFLAGS := -Wall -Wextra -Werror -I$(LIBFT_DIR) -I$(INC_DIR) -fsanitize=address #-g3   
-LDFLAGS := -L$(LIBFT_DIR) -fsanitize=address 
-LDLIBS := -lft
+CFLAGS := -Wall -Wextra -Werror -I$(LIBFT_DIR) -I$(INC_DIR) # -fsanitize=address #-g3   
+LDFLAGS := -L$(LIBFT_DIR) 
+LDLIBS := -lft -lm
 
 
 SRCS := ft_ping.c \
@@ -21,20 +21,13 @@ SRCS := ft_ping.c \
 	options_check.c \
 	options_error.c \
 	session.c \
+	sequence.c \
 	host.c \
 	socket.c \
 	icmp_datagram.c \
 	packet.c \
 	time.c
-	# verify_header.c \
-	# receive.c \
-	# print.c \
-	# data.c \
-	# signal_handler.c \
-	# get_time.c \
-	# host.c \
-	# socket.c \
-	# loop.c
+
 OBJS := $(SRCS:%.c=%.o)
 
 SRC_PATH := $(addprefix $(SRC_DIR)/,$(SRCS))
